refactor: SectionResult 패키지 리팩토링
