+++ src/test/java/wooteco/subway/domain/SectionTest.java
@@ -2,6 +2,7 @@ package wooteco.subway.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import wooteco.subway.dto.SectionResult;
 
 import java.util.List;
